break overview into it's own component
