+++ src/components/SimulationReport/simulation-report-view.vue
@@ -33,113 +33,8 @@
           class="overview-card"
           no-gutters
         >
-          <div id="overview-top" />
-          <!-- Overview -->
-          <v-col>
-            <v-row
-              ref="overviewScroll"
-              no-gutters
-              class="overview-list-container"
-            >
-              <div
-                v-if="overviewArrow && !scrollArrowConsumed"
-                class="align-center d-flex justify-center scroll-arrow"
-                @click="scrollToEndOfOverview"
-              >
-                <v-icon
-                  size="large"
-                  style="z-index:102; position:absolute;"
-                >
-                  mdi-circle
-                </v-icon>
-                <v-icon
-                  color="white"
-                  style="z-index:103;"
-                >
-                  mdi-chevron-double-right
-                </v-icon>
-              </div>
-
-              <v-col
-                v-for="{report, printjob, workflow, color} in reportData"
-                :key="report.id"
-                class="overview-list"
-                style=""
-              >
-                <!-- Data -->
-                <v-row
-                  style="overflow:hidden; height:100%;"
-                  no-gutters
-                >
-                  <v-col style="width:12vw; height:100%; max-height:100%;">
-                    <!-- Simulation Report Title -->
-                    <div
-                      class="overview-list-header"
-                    >
-                      <v-chip
-                        :color="color"
-                        size="x-small"
-                      >
-                        {{ printjob.Title }}
-                      </v-chip>
-                      <br />
-                      <v-chip
-                        size="x-small"
-                        :color="color"
-                      >
-                        {{ workflow.Title }}
-                      </v-chip>
-                    </div>
-
-                    <!-- Simulation Report Labels -->
-                    <div
-                      v-for="header in headersOverview.report"
-                      :key="header.value"
-                      class="overview-list-item"
-                    >
-                      <div class="overview-list-label">
-                        {{ header.text }}
-                      </div>
-                      <div class="overview-list-value">
-                        {{ report[header.value] }}
-                      </div>
-                      <br />
-                    </div>
-
-                    <!-- Print Job Labels -->
-                    <div
-                      v-for="header in headersOverview.printjob"
-                      :key="header.value"
-                      class="overview-list-item"
-                    >
-                      <div class="overview-list-label">
-                        {{ header.text }}
-                      </div>
-                      <div class="overview-list-value">
-                        {{ printjob[header.value] }}
-                      </div>
-                      <br />
-                    </div>
-
-                    <!-- Workflow Labels -->
-                    <div
-                      v-for="header in headersOverview.workflow"
-                      :key="header.value"
-                      class="overview-list-item"
-                    >
-                      <div class="overview-list-label">
-                        {{ header.text }}
-                      </div>
-                      <div class="overview-list-value">
-                        {{ workflow[header.value] }}
-                      </div>
-                      <br />
-                    </div>
-                  </v-col>
-                </v-row>
-              </v-col>
-            </v-row>
-          </v-col>
+          <report-overview :report-data="reportData">
+          </report-overview>
         </v-row>
         <!-- End Overview -->
 
@@ -461,9 +356,10 @@
 </template>
 
 <script setup>
-import { onMounted, ref, computed, nextTick, reactive } from "vue";
+import { onMounted, ref, computed, nextTick } from "vue";
 import {getPrintJob, getWorkflow, getCollection, getWorkflowTimes } from "../api.js";
 import ChartAll from '../Chart/chart-all.vue';
+import ReportOverview from './View/overview.vue';
 
 //// Props
 const {
@@ -496,31 +392,14 @@ const selectedChart = ref(null);
 const labels = ref([]);
 const chartCanvas = ref(null);
 
-////
-const headersOverview = ref({
-  report: [
-    { text: "Date Created", value: "dateTime" },
-    { text: "Total Time (s)", value: "TotalTimeTaken" },
-    { text: "Rasterization Profile", value: "RasterizationProfile" },
-  ],
-  printjob: [
-    { text: "Page Count", value: "PageCount" },
-  ],
-  workflow: [
-  ],
-});
-
 //// Scroll Data
 const comparisonScroll = ref(null);
 const comparisonArrow = ref(false);
-const overviewScroll = ref(null);
 const overviewArrow = ref(false);
 const scrollArrowConsumed = ref(false);
 const checkScroll = () => {
   const element = comparisonScroll.value.$el;
   comparisonArrow.value = element.scrollLeft + element.clientWidth < element.scrollWidth;
-  const element2 = overviewScroll.value.$el;
-  overviewArrow.value = element2.scrollLeft + element2.clientWidth < element2.scrollWidth;
 }
 
 //////////////////
@@ -549,20 +428,6 @@ const selectedChartLabels = computed(() => {
 
 const scrollToBottomOfCharts = () => {
   document.getElementById("chart-bottom").scrollIntoView();
-};
-
-const scrollToTopOfOverview = () => {
-  document.getElementById("overview-top").scrollIntoView();
-};
-
-const scrollToEndOfOverview = () => {
-  const refval = overviewScroll.value;
-  if (refval) {
-    const el = refval.$el;
-    const max = el.scrollWidth;
-    el.scrollBy(max,0, 'smooth');
-  }
-  scrollArrowConsumed.value = true;
 };
 
 const scrollToTopOfComparison = () => {
@@ -683,21 +548,6 @@ const highlightStep = (index) => {
   }
 }
 
-/**
-* Calculates how wide each report in the overview should be
-* Based on the assumption each of the two sections (data and graph)
-* takes up roughly have the viewport.
-*/
-const calculateOverviewWidth = () => {
-  const reportCount = reportData.value.length ? reportData.value.length : 1;
-  if (reportCount === 1) {
-    return '100%';
-  }
-  if (reportCount <= 3) {
-    return (45 / reportCount) + 'vw';
-  }
-  return (45 / 3) + 'vw';
-};
 
 ///////////////////////
 //// Logic
@@ -970,6 +820,7 @@ onMounted(
 .overview-card {
   display:block;
   height: calc((var(--section-height) - var(--vertical-gap)) / 2);
+  justify-content:start;
 
   max-width:inherit;
   min-width:inherit;
@@ -1053,39 +904,6 @@ onMounted(
   box-shadow:none;
 }
 
-.overview-container{
-  margin-left:20px;
-  margin-right:20px;
-  margin-bottom:20px;
-  overflow-x:hidden;
-  overflow-y:hidden;
-  border-style:solid;
-  border-width:1px;
-  border-radius:0px;
-  border-color:rgba(0,0,0,0.2);
-}
-
-.overview-item{
-  border-width:1px;
-  margin:0;
-  padding:0;
-}
-
-.overview-card-title {
-  font-size: 1.2em;
-  height:18px;
-  text-align:center;
-  text-justify: center;
-  background-color:rgb(25,25,25);
-  color:white;
-  font-weight: bold;
-  font-size: 1.0em;
-  font-family: 'Courier New', Courier, monospace;
-  border-radius:0px;
-  text-transform: uppercase;
-  margin-bottom:10px;
-
-}
 
 .comparison-card-title{
   height:18px;
@@ -1153,75 +971,6 @@ onMounted(
  75%   { scale: 2.0; }
  90%   { scale: 2.0; opacity:1; }
  100%  { scale: 2.0; opacity:0; }
-}
-
-.overview-list-item{
-  width:100%;
-  height:14.28%;
-  border-width:1px;
-}
-
-.overview-list-header{
-  width:100%;
-  text-align:center;
-  align-content:center;
-  height:20%;
-  color:white;
-  font-family:'Courier New', Courier;
-  text-overflow:ellipsis;
-  overflow:hidden;
-  text-wrap:nowrap;
-  max-height:40%;
-}
-
-.overview-list-label{
-  opacity: 0.7;
-  text-overflow:hidden;
-  text-align: center;
-  max-height:40%;
-  max-width:80%;
-  margin-left:auto;
-  margin-right:auto;
-  text-wrap:nowrap;
-  overflow:hidden;
-  font-size:0.8em;
-}
-
-.overview-list-value{
-  opacity:1.0;
-  text-overflow:ellipsis;
-  max-height:60%;
-  max-width:80%;
-  margin-left:auto;
-  margin-right:auto;
-  text-align: center;
-  text-wrap:nowrap;
-  overflow:hidden;
-  font-size:0.9em;
-}
-
-.overview-list-container {
-  height:100%;
-  position:relative;
-  display:flex;
-  flex-direction: row;
-  overflow-x:scroll;
-  overflow-y:hidden;
-  flex-wrap:nowrap;
-  flex-grow:1;
-}
-
-.overview-list{
-  background:white;
-  height:inherit;
-  max-height:inherit;
-  width:14vw;
-  max-width:14vw;
-  border-width:1px;
-  border-color:rgba(50,50,50,0.5);
-  margin-right:5px;
-  border-radius: 15px;
-  border-style:groove;
 }
 
 .scroll-arrow {
